requirements.txt
# Universal Insurance AI Agent - Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Date/Time handling
python-dateutil>=2.8.2

# Authentication & Security
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing
bcrypt>=4.0.0

# Database
sqlalchemy>=2.0.0  # ORM
asyncpg>=0.29.0  # Async PostgreSQL driver
alembic>=1.13.0  # Database migrations
psycopg2-binary>=2.9.9  # PostgreSQL adapter (sync)

# PDF Processing & OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0  # PaddlePaddle backend for PaddleOCR
pdf2image>=1.16.0
Pillow>=10.0.0
PyMuPDF>=1.23.0  # For PDF manipulation (fitz)
numpy>=1.24.0

# Vector Store & Embeddings
sentence-transformers>=2.2.0  # For generating embeddings (fallback)
chromadb>=0.4.0  # Lightweight vector database (optional)
scikit-learn>=1.3.0  # For cosine similarity calculations
pgvector>=0.2.4  # PostgreSQL vector extension for persistent storage
asyncpg>=0.29.0  # Async PostgreSQL driver

# Enhanced RAG
rank-bm25>=0.2.2  # BM25 for keyword search
nltk>=3.8.0  # Text processing for better chunking

# Agent Framework (Reasoning Loop)
langgraph>=1.0.0  # LangGraph for agent orchestration
langchain-core>=0.2.0  # Core abstractions

# LLM Providers (optional - for production)
openai>=1.6.0  # For GPT-4o integration (+ embeddings API)
anthropic>=0.18.0  # For Claude integration
google-generativeai>=0.3.0  # For Google Gemini integration
tiktoken>=0.5.0  # Token counting for OpenAI

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-html>=4.1.0
pytest-json-report>=1.5.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For async test client
